+++ precompile/erc721/Cargo.toml
@@ -34,6 +34,7 @@ num_enum = { workspace = true }
 [dev-dependencies]
 evm = { workspace = true }
 hex = { version = "0.4.3" }
+precompile-utils = { workspace = true, features = ["testing"]}
 
 [features]
 default = ["std"]
