+++ precompile/living-assets/Cargo.toml
@@ -38,6 +38,7 @@ num_enum = { workspace = true }
 [dev-dependencies]
 evm = { workspace = true }
 hex = { version = "0.4.3" }
+precompile-utils = { workspace = true, features = ["testing"]}
 
 [features]
 default = ["std"]
