+++ precompile/utils/src/testing.rs
@@ -224,6 +224,18 @@ impl PrecompileHandle for MockHandle {
 	fn gas_limit(&self) -> Option<u64> {
 		Some(self.gas_limit)
 	}
+
+	fn record_external_cost(
+		&mut self,
+		_ref_time: Option<u64>,
+		_proof_size: Option<u64>,
+	) -> Result<(), ExitError> {
+		todo!()
+	}
+
+	fn refund_external_cost(&mut self, _ref_time: Option<u64>, _proof_size: Option<u64>) {
+		todo!()
+	}
 }
 
 pub struct PrecompilesTester<'p, P> {
